Ignore check prefixes at the ends of words
Ignore check prefixes that are actually the ends of longer words. For example, a
line containing "NO-CHECK:" should not be interpreted as containing a "CHECK:"
directive. A check prefix is considered to be the end of a longer word if there
is a character before it on the same line and that character is alphanumeric,
"_", or "-", just like in upstream FileCheck.